Rework when compound statements are used
Avoids the use of lists of statements in favour of compound statements
to be more in line with WGSL syntax, and to allow the treatment of
loop bodies, function bodies and continue statement bodies in a manner
that is consistent with if, for and while bodies.

Getting this right involves some changes to the resolver to ensure
that scoping is respected.

Fixes #15.