Prototype dataflow-based uniformity analysis
Adds a prototype implementation of a dataflow-based uniformity
analysis. The motivation for this is to help in studying the
graph-based uniformity analysis described in the WGSL specification
and implemented in Tint. The analysis contributed here makes very
strong assumptions about the structure of the input WGSL program -
e.g., just one function, no real expressions (only identifiers). These
restrictions could be lifted to enable a more realistic comparison
with other analysis implementations.