Start defining featherweight fragment of WGSL
To help in experimenting with different formulations of uniformity
analysis it would be useful to pin down a minimal fragment of WGSL
that allows one to write compute shaders with interesting control
flow, but where any features of the language superflous to the essence
of uniformity analysis are banned. This change makes a start on
defining this fragment by introducing checks on a WGSL translation
unit that will reject shaders that use all but minimal features. The
existing uniformity data flow analysis is updated accordingly.