Overhaul the way scopes are handled
A reworking of parts of the resolver so that it is possible to query,
via a scope, what is available right before each statement and right
at the end of every compound statement. This is a precursor to
enabling the generation of randomized expressions that make use of
in-scope declarations at a particular program point.