Add node limit to generator, plus misc fixes
Adds a configurable AST node limit to the generator. Moves logic that
checks every node in the AST is typed so that this only runs in the
TransformReduceTests, fixes a bug in that logic, and fixes a problem
subsequently exposed by those tests wereby type declarations in
parameters were not being resolved.