Check if wrapping ControlFlowWrapReturn (#231)
Originally if a ControlFlowWrapReturn was wrapped another return would
not be inserted at the end of the compound to ensure that every code
path had a return. This meant transformed programs may not of been
valid.

The major change is to do with the traversal state storing the
return type decl for the function it is in. This can then be used to
generate an arbitrary expression of the correct type.

Fixes: #230